Move go project test
