+++ test/com/facebook/buck/features/project/intellij/testdata/go_project_with_existing_vendor/.buckconfig
