+++ test/com/facebook/buck/features/project/intellij/testdata/go_project_with_generated_code/BUCK.fixture
