+++ test/com/facebook/buck/features/project/intellij/testdata/go_project_with_generated_code/a/b1.go
@@ -1,0 +1,3 @@
+package a
+func A() {}
+
